@@ -10,7 +10,7 @@
     "check-types": "tsc",
     "dev": "next dev",
     "export": "next export",
-    "lint": "eslint --ext ts,tsx .",
+    "lint": "eslint --ext ts,tsx --max-warnings=0 .",
     "start": "next start",
     "test": "is-ci-cli \"test:coverage\" \"test:watch\"",
     "test:watch": "node --experimental-vm-modules node_modules/jest/bin/jest.js --watch",
@@ -21,21 +21,18 @@
   },
   "prettier": "@lynxtaa/prettier-config",
   "dependencies": {
-    "@next/bundle-analyzer": "13.0.5",
-    "@tanstack/react-query": "4.18.0",
+    "@next/bundle-analyzer": "13.2.0",
+    "@tanstack/react-query": "4.24.4",
     "autoprefixer": "10.4.13",
     "cheerio": "1.0.0-rc.12",
     "classnames": "2.3.2",
     "date-fns": "2.29.3",
     "date-fns-tz": "1.3.7",
     "focus-visible": "5.2.0",
-    "got": "11.8.5",
     "html-react-parser": "1.4.14",
     "lodash": "4.17.21",
-    "next": "13.0.5",
-    "next-seo": "5.14.1",
+    "next": "13.2.0",
     "next-themes": "0.2.1",
-    "nprogress": "0.2.0",
     "p-queue": "6.6.2",
     "postcss": "8.4.19",
     "postcss-focus-visible": "7.1.0",
@@ -44,10 +41,11 @@
     "react-dom": "18.2.0",
     "react-feather": "2.0.10",
     "rss": "1.2.2",
+    "server-only": "^0.0.1",
     "sharp": "0.31.2",
     "tailwind-override": "0.6.1",
     "tailwindcss": "3.2.4",
-    "undici": "5.13.0",
+    "undici": "5.20.0",
     "zod": "3.19.1"
   },
   "devDependencies": {
@@ -56,27 +54,27 @@
     "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "13.4.0",
     "@testing-library/user-event": "14.4.3",
-    "@types/jest": "29.2.3",
-    "@types/lodash": "4.14.190",
-    "@types/node": "16.18.3",
-    "@types/nprogress": "0.2.0",
+    "@types/jest": "29.4.0",
+    "@types/lodash": "4.14.191",
+    "@types/node": "16.18.12",
     "@types/probe-image-size": "7.2.0",
-    "@types/react": "18.0.25",
-    "@types/react-dom": "18.0.9",
+    "@types/react": "18.0.28",
+    "@types/react-dom": "18.0.10",
     "@types/rss": "0.0.29",
     "cross-env": "7.0.3",
-    "eslint": "8.28.0",
-    "eslint-config-next": "13.0.5",
+    "eslint": "8.34.0",
+    "eslint-config-next": "13.2.0",
     "eslint-plugin-react-hooks": "4.6.0",
     "is-ci-cli": "2.2.0",
-    "jest": "29.3.1",
+    "jest": "29.4.2",
     "jest-date-mock": "1.0.8",
-    "jest-environment-jsdom": "29.3.1",
-    "msw": "0.49.0",
-    "prettier": "2.8.0",
+    "jest-environment-jsdom": "29.4.2",
+    "msw": "1.1.0",
+    "prettier": "2.8.4",
     "pretty-quick": "3.1.3",
     "simple-git-hooks": "2.8.1",
-    "typescript": "4.9.3",
+    "typescript": "4.9.5",
     "whatwg-fetch": "3.6.2"
-  }
+  },
+  "packageManager": "pnpm@7.27.0"
 }
