@@ -1,28 +1,17 @@
-import { NextRouter } from 'next/router'
+import { useRouter } from 'next/navigation'
 
-export function createMockRouter(options?: Partial<NextRouter>): NextRouter {
+export type AppRouterInstance = ReturnType<typeof useRouter>
+
+export function createMockRouter(
+	options?: Partial<AppRouterInstance>,
+): AppRouterInstance {
 	return {
-		basePath: '/',
-		pathname: '/',
-		route: '/',
-		query: {},
-		asPath: '/',
-		push: jest.fn(() => Promise.resolve(true)),
-		replace: jest.fn(() => Promise.resolve(true)),
-		reload: jest.fn(),
-		prefetch: jest.fn(() => Promise.resolve()),
+		push: jest.fn(),
+		replace: jest.fn(),
+		refresh: jest.fn(),
+		prefetch: jest.fn(),
 		back: jest.fn(),
 		forward: jest.fn(),
-		beforePopState: jest.fn(),
-		isFallback: false,
-		isLocaleDomain: true,
-		events: {
-			on: jest.fn(),
-			off: jest.fn(),
-			emit: jest.fn(),
-		},
-		isReady: true,
-		isPreview: false,
 		...options,
 	}
 }
